♻️refactor: UserVectorVO를 userPolicy 패키지로 이동 및 구 파일 정리 #35
